requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
cachetools==5.3.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap3==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-filter==22.1
djangorestframework==3.14.0
grpcio==1.51.3
isort==5.12.0
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
plotly==5.13.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2-binary==2.9.5
pydantic==1.10.6
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
tinkoff==0.1.1
tinkoff-investments==0.2.0b50
tomli==2.0.1
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
